+++ src/common/common.py
@@ -16,7 +16,7 @@ from gql.transport.exceptions import TransportQueryError
 from .api_requester import GQLRequester
 from ..metadata import RunnerMetadata, Block
 from .. import config
-from ... import cache_scripts
+from dao_analyzer import cache_scripts
 
 # To be able to obtain endpoints.json
 ENDPOINTS: Dict = json.loads(pkgutil.get_data(cache_scripts.__name__, 'endpoints.json'))