+++ test/daohaus/test_verify_daohaus_collectors.py
@@ -1,0 +1,9 @@
+import pytest
+
+from src.common.graphql import GraphQLCollector
+from src.daohaus.runner import DaohausRunner
+
+@pytest.mark.parametrize("c", DaohausRunner().collectors)
+def test_verify_collectors(c):
+    assert c.verify()
+